+++ app/Http/Controllers/Server/ServerController.php
@@ -239,15 +239,14 @@ class ServerController extends Controller
     {
         $server = Models\Server::byUuid($uuid);
         $this->authorize('view-databases', $server);
+
+        $server->load('node', 'databases.host');
         $server->js();
 
         return view('server.settings.databases', [
             'server' => $server,
             'node' => $server->node,
-            'databases' => Models\Database::select('databases.*', 'database_servers.host as a_host', 'database_servers.port as a_port')
-                ->where('server_id', $server->id)
-                ->join('database_servers', 'database_servers.id', '=', 'databases.db_server')
-                ->get(),
+            'databases' => $server->databases,
         ]);
     }